Changed extension names to cpp
