+++ cpp-RayTracer/src/hittable.hpp
