+++ cpp-RayTracer/src/interval.hpp
