+++ cpp-RayTracer/src/material.hpp
@@ -1,7 +1,7 @@
 #ifndef MATERIAL_H
 #define MATERIAL_H
 
-#include "hittable.h"
+#include "hittable.hpp"
 
 class material {
   public:
