+++ cpp-RayTracer/src/ray.hpp
@@ -1,7 +1,7 @@
 #ifndef RAY_H
 #define RAY_H
 
-#include "vec3.h"
+#include "vec3.hpp"
 
 class ray {
   public:
