+++ cpp-RayTracer/src/vec3.hpp
